+++ composer.json
@@ -1,42 +1,41 @@
 {
-    "name": "appserver-io/description",
-    "description": "Deployment Descriptor implementation used by appserver.io.",
-    "homepage": "https://github.com/appserver-io/description",
-    "license": "OSL-3.0",
-    "require": {
-        "php": ">=5.4.0",
-        "rhumsaa/uuid": "~2.4"
-    },
-    "require-dev": {
-        "appserver-io/build": "~1.0",
-        "appserver-io/properties": "~2.0",
-        "appserver-io/configuration": "~2.0",
-        "appserver-io-psr/epb":"dev-di as 4.0.0",
-        "appserver-io-psr/servlet": "dev-di as 3.1.0",
-        "appserver-io-psr/deployment": "~1.0"
-    },
-    "autoload": {
-        "psr-0": {
-            "AppserverIo\\Description": [
-                "src/",
-                "tests/"
-            ]
-        }
-    },
-    "authors": [
-        {
-            "name": "Tim Wagner",
-            "email": "tw@appserver.io",
-            "homepage": "https://github.com/wagnert",
-            "role": "Developer"
-        }
-    ],
-    "support": {
-        "email": "t.wagner@techdivision.com",
-        "issues": "https://github.com/appserver-io/description/issues",
-        "source": "https://github.com/appserver-io/description"
-    },
-    "keywords": [
-        "deployment descriptor implementation"
-    ]
+	"name" : "appserver-io/description",
+	"description" : "Deployment Descriptor implementation used by appserver.io.",
+	"homepage" : "https://github.com/appserver-io/description",
+	"license" : "OSL-3.0",
+	"require" : {
+		"php" : ">=5.4.0",
+		"rhumsaa/uuid" : "~2.4"
+	},
+	"require-dev" : {
+		"appserver-io/build" : "~1.0",
+		"appserver-io/properties" : "~1.0",
+		"appserver-io/configuration" : "~2.0",
+		"appserver-io-psr/epb" : "~4.0",
+		"appserver-io-psr/servlet" : "~1.0",
+		"appserver-io-psr/deployment" : "~1.0"
+	},
+	"autoload" : {
+		"psr-0" : {
+			"AppserverIo\\Description" : [
+				"src/",
+				"tests/"
+			]
+		}
+	},
+	"authors" : [{
+			"name" : "Tim Wagner",
+			"email" : "tw@appserver.io",
+			"homepage" : "https://github.com/wagnert",
+			"role" : "Developer"
+		}
+	],
+	"support" : {
+		"email" : "t.wagner@techdivision.com",
+		"issues" : "https://github.com/appserver-io/description/issues",
+		"source" : "https://github.com/appserver-io/description"
+	},
+	"keywords" : [
+		"deployment descriptor implementation"
+	]
 }
