+++ src/AppserverIo/Description/Api/Node/MessageDrivenNode.php
@@ -67,14 +67,6 @@ class MessageDrivenNode extends AbstractNode implements MessageDrivenConfigurati
     protected $resRefs = array();
 
     /**
-     * The bean reference information.
-     *
-     * @var array
-     * @AS\Mapping(nodeName="bean-ref", nodeType="array", elementType="AppserverIo\Description\Api\Node\BeanRefNode")
-     */
-    protected $beanRefs = array();
-
-    /**
      * The persistence unit reference information.
      *
      * @var array
@@ -120,16 +112,6 @@ class MessageDrivenNode extends AbstractNode implements MessageDrivenConfigurati
     public function getResRefs()
     {
         return $this->resRefs;
-    }
-
-    /**
-     * Return's the bean reference information.
-     *
-     * @return array The bean reference information
-     */
-    public function getBeanRefs()
-    {
-        return $this->beanRefs;
     }
 
     /**
