+++ src/AppserverIo/Description/Api/Node/ServletNode.php
@@ -91,14 +91,6 @@ class ServletNode extends AbstractNode implements ServletConfigurationInterface
     protected $resRefs = array();
 
     /**
-     * The class reference information.
-     *
-     * @var array
-     * @AS\Mapping(nodeName="class-ref", nodeType="array", elementType="AppserverIo\Description\Api\Node\ClassRefNode")
-     */
-    protected $classRefs = array();
-
-    /**
      * The persistence unit reference information.
      *
      * @var array
@@ -174,16 +166,6 @@ class ServletNode extends AbstractNode implements ServletConfigurationInterface
     public function getResRefs()
     {
         return $this->resRefs;
-    }
-
-    /**
-     * Return's the class reference information.
-     *
-     * @return array The class reference information
-     */
-    public function getBeanRefs()
-    {
-        return $this->classRefs;
     }
 
     /**
