+++ src/AppserverIo/Description/Api/Node/SessionNode.php
@@ -155,14 +155,6 @@ class SessionNode extends AbstractNode implements SessionConfigurationInterface
     protected $resRefs = array();
 
     /**
-     * The bean reference information.
-     *
-     * @var array
-     * @AS\Mapping(nodeName="bean-ref", nodeType="array", elementType="AppserverIo\Description\Api\Node\BeanRefNode")
-     */
-    protected $beanRefs = array();
-
-    /**
      * The persistence unit reference information.
      *
      * @var array
@@ -318,16 +310,6 @@ class SessionNode extends AbstractNode implements SessionConfigurationInterface
     public function getResRefs()
     {
         return $this->resRefs;
-    }
-
-    /**
-     * Return's the bean reference information.
-     *
-     * @return array The bean reference information
-     */
-    public function getBeanRefs()
-    {
-        return $this->beanRefs;
     }
 
     /**
