--- src/AppserverIo/Description/EnterpriseBeanDescriptor.php
@@ -1,79 +1,0 @@
-<?php
-
-/**
- * AppserverIo\Description\EnterpriseBeanDescriptor
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is available through the world-wide-web at this URL:
- * http://opensource.org/licenses/osl-3.0.php
- *
- * PHP version 5
- *
- * @author    Tim Wagner <tw@appserver.io>
- * @copyright 2015 TechDivision GmbH <info@appserver.io>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link      https://github.com/appserver-io/description
- * @link      http://www.appserver.io
- */
-
-namespace AppserverIo\Description;
-
-use AppserverIo\Lang\Reflection\ClassInterface;
-
-/**
- * Abstract class for all enterprise bean descriptors.
- *
- * @author    Tim Wagner <tw@appserver.io>
- * @copyright 2015 TechDivision GmbH <info@appserver.io>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link      https://github.com/appserver-io/description
- * @link      http://www.appserver.io
- */
-abstract class EnterpriseBeanDescriptor extends BeanDescriptor
-{
-
-    /**
-     * Returns a new annotation instance for the passed reflection class.
-     *
-     * @param \AppserverIo\Lang\Reflection\ClassInterface $reflectionClass The reflection class with the bean configuration
-     *
-     * @return \AppserverIo\Lang\Reflection\AnnotationInterface The reflection annotation
-     */
-    abstract protected function newAnnotationInstance(ClassInterface $reflectionClass);
-
-    /**
-     * Initializes the bean configuration instance from the passed reflection class instance.
-     *
-     * @param \AppserverIo\Lang\Reflection\ClassInterface $reflectionClass The reflection class with the bean configuration
-     *
-     * @return void
-     */
-    public function fromReflectionClass(ClassInterface $reflectionClass)
-    {
-
-        // create a new annotation instance
-        $reflectionAnnotation = $this->newAnnotationInstance($reflectionClass);
-
-        // load class name
-        $this->setClassName($reflectionClass->getName());
-
-        // initialize the annotation instance
-        $annotationInstance = $reflectionAnnotation->newInstance(
-            $reflectionAnnotation->getAnnotationName(),
-            $reflectionAnnotation->getValues()
-        );
-
-        // load the default name to register in naming directory
-        if ($nameAttribute = $annotationInstance->getName()) {
-            $this->setName(DescriptorUtil::trim($nameAttribute));
-        } else {
-            // if @Annotation(name=****) is NOT set, we use the short class name by default
-            $this->setName($reflectionClass->getShortName());
-        }
-
-        // initialize references from the passed reflection class
-        $this->referencesFromReflectionClass($reflectionClass);
-    }
-}
