+++ src/AppserverIo/Description/SingletonSessionBeanDescriptor.php
@@ -173,7 +173,8 @@ class SingletonSessionBeanDescriptor extends SessionBeanDescriptor implements Si
     /**
      * Returns a new descriptor instance.
      *
-     * @return \AppserverIo\Psr\EnterpriseBeans\Description\SingletonSessionBeanDescriptorInterface The descriptor instance
+     * @return \AppserverIo\Psr\EnterpriseBeans\Description\SingletonSessionBeanDescriptorInterface
+     *     The descriptor instance
      */
     public static function newDescriptorInstance()
     {
